Force Usage of custom Reader in Test
