--- src/test/groovy/com/cedarsoftware/util/io/CustomDataClass.groovy
@@ -1,15 +1,0 @@
-package com.cedarsoftware.util.io
-
-class CustomDataClass
-{
-
-	private String test;
-
-	public void setTest(String test) {
-		this.test = test;
-	}
-
-	public String getTest() {
-		return test;
-	}
-}
